+++ app/main/forms.py
@@ -1,0 +1,7 @@
+from flask_wtf import FlaskForm
+from wtforms import StringField
+from wtforms.validators import DataRequired
+
+
+class MovieTitleForm(FlaskForm):
+    title = StringField("Movie Title", validators=[DataRequired()])
