+++ app/main/models.py
@@ -1,0 +1,12 @@
+from app import db
+
+
+class Movie(db.Model):
+    id = db.Column(db.Integer, primary_key=True)
+    title = db.Column(db.String, unique=True)
+    year = db.Column(db.Integer)
+    description = db.Column(db.String)
+    rating = db.Column(db.Float)
+    ranking = db.Column(db.Integer)
+    review = db.Column(db.String)
+    img_url = db.Column(db.String)
