+++ app/templates/index.html
@@ -21,8 +21,8 @@
           <p class="review">{{ movie.review }}</p>
         <p class="overview">{{ movie.description }}</p>
 
-        <a href="{{ url_for('edit', id=movie.id) }}" class="button">Update</a>
-        <a href="{{ url_for('delete', id=movie.id) }}" class="button delete-button">Delete</a>
+        <a href="{{ url_for('main.edit', id=movie.id) }}" class="button">Update</a>
+        <a href="{{ url_for('main.delete', id=movie.id) }}" class="button delete-button">Delete</a>
 
       </div>
     </div>
@@ -30,7 +30,7 @@
     {% endfor %}
 </div>
 <div class="container text-center add">
-<a href="{{ url_for('add') }}" class="button">Add Movie</a>
+<a href="{{ url_for('main.add') }}" class="button">Add Movie</a>
 </div>
 
 {% endblock %}
