+++ app/templates/select.html
@@ -9,7 +9,7 @@
       {% for item in data %}
       <li style="list-style-type: none">
         <ul>
-        <a href="{{ url_for('find_movie', id=item['id']) }}"> {{ item['title'] }} - {{ item['release_date'] }}</a>
+        <a href="{{ url_for('main.find_movie', id=item['id']) }}"> {{ item['title'] }} - {{ item['release_date'] }}</a>
         </ul>
       </li>
       {% endfor %}
